Wrong tileset is referenced when using multiple tilesets
Tiled uses the tileset-local tile id for animations rather than the tile gid.
This is a problem because tileset tile id's overlap when using multiple tilesets
with animations.

This PR adds a test for demonstrating the wrong tileset is chosen as well as a
fix for it.

This changes the tile gid to be globally unique based on the first gid of each tileset.
Map::processData is updated so the tilemap maps tile gid -> tile instead of tile id -> tile.
This allows the correct tile and tileset to be surfaced.