Removed "version" from tson::Map to support Tiled v1.6 #37
TiledVersion is still an own thing, which is still contained inside the map. Reason why version (json format version) was removed is that Tiled v1.6 changed the type from float to string, and it seems to give no real benefit to include in a map, so the easiest and most stable solution is to remove.