Fix user overrides of class attributes
When setting a class as property of another class, the attributes
are inherited and an override can be set for a attribute.

Previously when any override was set, the class was being replaced
by an empty class because it could not be found in the project,
because the reference to the project was not present in the add
method.

The reference was now added with an extra method to get the Project
pointer to make the compiler happy when building the amalgamation.