Add getFullImagePath(), returning a path relative to the parent path
Currently getImagePath() returns a path relative to itself, but
that is not very useful if the Tileset is loaded by querying it from a Map.

getFullImagePath() returns a path relative to the parent, that
can be resolved to the actual image file without scanning the
directories and looking for a file of similar name.

If the parent path doesn't exist, it will return the relative
path to itself, which is the current behavior.